# openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cave Daemon API",
    "version": "0.1.0",
    "description": "REST interface for managing sandboxes, executing workloads, and issuing API keys. See docs/api.md for narrative documentation."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List registered models",
        "operationId": "listModels",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Registered models",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelRecord"
                  }
                },
                "example": [
                  {
                    "id": "f3f5b6c4-9a62-4d1a-9da3-21b9a5b27f41",
                    "name": "phi-3",
                    "provider": "open-model",
                    "version": "1.0.0",
                    "format": "gguf",
                    "source_uri": "https://example.com/models/phi-3.gguf",
                    "checksum_sha256": "d2c7e95e1a1d4cbb7e1e5b6f13fd8e05de0b4d8717b4c2a38a7f1bbf3b189e0c",
                    "size_bytes": 1572864000,
                    "stage": "registered",
                    "last_synced_at": null,
                    "created_at": "2025-10-19T09:15:00Z",
                    "updated_at": "2025-10-19T09:15:00Z",
                    "tags": [
                      "text",
                      "demo"
                    ],
                    "error_message": null
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "missing Authorization bearer token"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "insufficient permissions for requested scope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "unexpected internal error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Register a model",
        "operationId": "registerModel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModelRequest"
              },
              "example": {
                "name": "phi-3",
                "provider": "open-model",
                "version": "1.0.0",
                "format": "gguf",
                "source_uri": "https://example.com/models/phi-3.gguf",
                "tags": [
                  "text",
                  "demo"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Model registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRecord"
                },
                "example": {
                  "id": "f3f5b6c4-9a62-4d1a-9da3-21b9a5b27f41",
                  "name": "phi-3",
                  "provider": "open-model",
                  "version": "1.0.0",
                  "format": "gguf",
                  "source_uri": "https://example.com/models/phi-3.gguf",
                  "checksum_sha256": "d2c7e95e1a1d4cbb7e1e5b6f13fd8e05de0b4d8717b4c2a38a7f1bbf3b189e0c",
                  "size_bytes": 1572864000,
                  "stage": "registered",
                  "last_synced_at": null,
                  "created_at": "2025-10-19T09:15:00Z",
                  "updated_at": "2025-10-19T09:15:00Z",
                  "tags": [
                    "text",
                    "demo"
                  ],
                  "error_message": null
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "namespace query parameter is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "missing Authorization bearer token"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "insufficient permissions for requested scope"
                }
              }
            }
          },
          "409": {
            "description": "Model already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "model 'phi-3' version '1.0.0' already exists"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "unexpected internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{id}/refresh": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Queue model refresh",
        "operationId": "refreshModel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model refresh queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRecord"
                },
                "example": {
                  "id": "f3f5b6c4-9a62-4d1a-9da3-21b9a5b27f41",
                  "name": "phi-3",
                  "provider": "open-model",
                  "version": "1.0.0",
                  "format": "gguf",
                  "source_uri": "https://example.com/models/phi-3.gguf",
                  "checksum_sha256": "d2c7e95e1a1d4cbb7e1e5b6f13fd8e05de0b4d8717b4c2a38a7f1bbf3b189e0c",
                  "size_bytes": 1572864000,
                  "stage": "queued",
                  "last_synced_at": null,
                  "created_at": "2025-10-19T09:15:00Z",
                  "updated_at": "2025-10-19T09:15:00Z",
                  "tags": [
                    "text",
                    "demo"
                  ],
                  "error_message": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "missing Authorization bearer token"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "insufficient permissions for requested scope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "sandbox 9f9c... not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "unexpected internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{id}": {
      "delete": {
        "tags": [
          "Models"
        ],
        "summary": "Delete a model",
        "operationId": "deleteModel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Model deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "missing Authorization bearer token"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "insufficient permissions for requested scope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "sandbox 9f9c... not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "unexpected internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{id}/jobs": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List model download jobs",
        "operationId": "listModelJobs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelJob"
                  }
                },
                "example": [
                  {
                    "id": "9f5a64a8-6c70-4e55-8b77-55c932677d4f",
                    "model_id": "f3f5b6c4-9a62-4d1a-9da3-21b9a5b27f41",
                    "stage": "queued",
                    "progress": 0.25,
                    "started_at": "2025-10-19T09:16:00Z",
                    "finished_at": null,
                    "error_message": null
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "missing Authorization bearer token"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "insufficient permissions for requested scope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "sandbox 9f9c... not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "unexpected internal error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human readable error message."
          }
        },
        "example": {
          "error": "sandbox 123 not found"
        }
      },
...
      "ModelStage": {
        "type": "string",
        "enum": [
          "unknown",
          "registered",
          "queued",
          "downloading",
          "verifying",
          "ready",
          "failed"
        ],
        "description": "Lifecycle state for model registration and download jobs."
      },
      "RegisterModelRequest": {
        "type": "object",
        "required": [
          "name",
          "provider",
          "version",
          "format",
          "source_uri"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "description": "Distribution format (e.g. gguf, safetensors)."
          },
          "source_uri": {
            "type": "string",
            "format": "uri"
          },
          "checksum_sha256": {
            "type": "string",
            "description": "Optional SHA-256 checksum for integrity validation."
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Approximate download size in bytes."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional labels for filtering."
          }
        },
        "example": {
          "name": "phi-3",
          "provider": "open-model",
          "version": "1.0.0",
          "format": "gguf",
          "source_uri": "https://example.com/models/phi-3.gguf",
          "tags": [
            "text",
            "demo"
          ]
        }
      },
      "ModelRecord": {
        "type": "object",
        "required": [
          "id",
          "name",
          "provider",
          "version",
          "format",
          "source_uri",
          "stage",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "source_uri": {
            "type": "string",
            "format": "uri"
          },
          "checksum_sha256": {
            "type": "string",
            "nullable": true
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "stage": {
            "$ref": "#/components/schemas/ModelStage"
          },
          "last_synced_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "error_message": {
            "type": "string",
            "nullable": true
          }
        },
        "example": {
          "id": "f3f5b6c4-9a62-4d1a-9da3-21b9a5b27f41",
          "name": "phi-3",
          "provider": "open-model",
          "version": "1.0.0",
          "format": "gguf",
          "source_uri": "https://example.com/models/phi-3.gguf",
          "checksum_sha256": "d2c7e95e1a1d4cbb7e1e5b6f13fd8e05de0b4d8717b4c2a38a7f1bbf3b189e0c",
          "size_bytes": 1572864000,
          "stage": "registered",
          "last_synced_at": null,
          "created_at": "2025-10-19T09:15:00Z",
          "updated_at": "2025-10-19T09:15:00Z",
          "tags": [
            "text",
            "demo"
          ],
          "error_message": null
        }
      },
      "ModelJob": {
        "type": "object",
        "required": [
          "id",
          "model_id",
          "stage",
          "progress",
          "started_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "stage": {
            "$ref": "#/components/schemas/ModelStage"
          },
          "progress": {
            "type": "number",
            "format": "float",
            "description": "Download progress between 0 and 1."
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "error_message": {
            "type": "string",
            "nullable": true
          }
        },
        "example": {
          "id": "9f5a64a8-6c70-4e55-8b77-55c932677d4f",
          "model_id": "f3f5b6c4-9a62-4d1a-9da3-21b9a5b27f41",
          "stage": "queued",
          "progress": 0.25,
          "started_at": "2025-10-19T09:16:00Z",
          "finished_at": null,
          "error_message": null
        }
      },
...
    }
  }
}